requirements.txt
pandas
langdetect
numpy==1.26.4
transformers
torch
youtube-comment-downloader

unstructured
sentence-transformers
faiss-cpu

langchain
markdown-it-py
mdit_plain

tqdm

streamlit